implement for scala 2
