+++ core/shared/src/main/scala-2/caseapp/core/help/WithHelpCompanion.scala
@@ -1,6 +1,6 @@
 package caseapp.core.help
 
-import caseapp.{ExtraName, Group, Help, HelpMessage, Parser}
+import caseapp.{ExtraName, Group, HelpMessage, Parser, Recurse}
 import caseapp.core.parser.{Argument, NilParser, StandardArgument}
 import caseapp.core.{Arg, Error}
 import caseapp.core.parser.{EitherParser, RecursiveConsParser}
@@ -56,7 +56,7 @@ abstract class WithHelpCompanion {
 
     val p = usageArgument ::
       helpArgument ::
-      RecursiveConsParser(either, NilParser)
+      RecursiveConsParser(either, NilParser, Recurse())
 
     p.to[WithHelp[T]]
   }
