+++ core/shared/src/main/scala-2/caseapp/core/parser/HListParserBuilder.scala
@@ -16,7 +16,7 @@ sealed abstract class HListParserBuilder[
   -G <: HList,
   -H <: HList,
   -T <: HList,
-  R <: HList
+  -R <: HList
 ] {
   type P <: HList
   def apply(
@@ -26,7 +26,8 @@ sealed abstract class HListParserBuilder[
     helpMessages: M,
     groups: G,
     noHelp: H,
-    tags: T
+    tags: T,
+    recurse: R
   ): Parser.Aux[L, P]
 }
 
@@ -73,7 +74,7 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
     R <: HList,
     P0 <: HList
   ](
-    p: (() => D, N, V, M, G, H, T) => Parser.Aux[L, P0]
+    p: (() => D, N, V, M, G, H, T, R) => Parser.Aux[L, P0]
   ): Aux[L, D, N, V, M, G, H, T, R, P0] =
     new HListParserBuilder[L, D, N, V, M, G, H, T, R] {
       type P = P0
@@ -84,13 +85,14 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
         helpMessages: M,
         group: G,
         noHelp: H,
-        tags: T
+        tags: T,
+        recurse: R
       ) =
-        p(() => default, names, valueDescriptions, helpMessages, group, noHelp, tags)
+        p(() => default, names, valueDescriptions, helpMessages, group, noHelp, tags, recurse)
     }
 
   implicit val hnil: Aux[HNil, HNil, HNil, HNil, HNil, HNil, HNil, HNil, HNil, HNil] =
-    instance { (_, _, _, _, _, _, _) =>
+    instance { (_, _, _, _, _, _, _, _) =>
       NilParser
     }
 
@@ -124,7 +126,7 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
     None.type :: RT,
     Option[H] :: PT
   ] =
-    instance { (default0, names, valueDescriptions, helpMessages, groups, noHelp, tags) =>
+    instance { (default0, names, valueDescriptions, helpMessages, groups, noHelp, tags, recurse) =>
 
       val tailParser = tail.value(
         default0().tail,
@@ -133,7 +135,8 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
         helpMessages.tail,
         groups.tail,
         noHelp.tail,
-        tags.tail
+        tags.tail,
+        recurse.tail
       )
 
       val arg = Arg(
@@ -181,7 +184,7 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
     Some[Recurse] :: RT,
     headParser.value.D :: PT
   ] =
-    instance { (default0, names, valueDescriptions, helpMessages, groups, noHelp, tags) =>
+    instance { (default0, names, valueDescriptions, helpMessages, groups, noHelp, tags, recurse) =>
 
       val tailParser = tail(
         default0().tail,
@@ -190,10 +193,11 @@ object HListParserBuilder extends LowPriorityHListParserBuilder {
         helpMessages.tail,
         groups.tail,
         noHelp.tail,
-        tags.tail
+        tags.tail,
+        recurse.tail
       )
 
-      RecursiveConsParser(headParser.value, tailParser)
+      RecursiveConsParser(headParser.value, tailParser, recurse.head.get)
         .mapHead(field[K](_))
     }
 }
