+++ core/shared/src/main/scala-2/caseapp/core/parser/LowPriorityParserImplicits.scala
@@ -40,7 +40,8 @@ trait LowPriorityParserImplicits {
         helpMessages(),
         group(),
         noHelp(),
-        tags()
+        tags(),
+        recurse()
       )
       .map(gen.from)
       .withDefaultOrigin(typeable.describe)
