+++ core/shared/src/main/scala/caseapp/core/util/Formatter.scala
@@ -1,6 +1,6 @@
 package caseapp.core.util
 
-import caseapp.Name
+import caseapp.{Name, Recurse}
 
 abstract class Formatter[T] {
   def format(t: T): String
@@ -19,4 +19,19 @@ object Formatter {
         .map(_.toLowerCase)
         .mkString("-")
   }
+
+  /** Adds the prefix for the Recurse annotation to the names formatted by the formatter.
+    *
+    * Adds the prefix as `prefix-foo-bar`.
+    */
+  def addRecursePrefix(recurse: Recurse, formatter: Formatter[Name]): Formatter[Name] =
+    if (recurse.prefix.isEmpty()) formatter
+    else {
+      new Formatter[Name] {
+        def format(t: Name): String = {
+          val formattedPrefix = formatter.format(Name(recurse.prefix))
+          s"${formattedPrefix}-${formatter.format(t)}"
+        }
+      }
+    }
 }
